Updated Entities, Repositories, and Controllers

+++ PizzaBox.Domain/Models/APizzaTopping.cs
@@ -1,13 +1,13 @@
 namespace PizzaBox.Domain.Models
 {
-  public partial class APizzaTopping
-  {
-    public int PizzaToppingId { get; set; }
-    public int PizzaId { get; set; }
-    public byte ToppingId { get; set; }
-    public byte ToppingCount { get; set; }
+    public partial class APizzaTopping
+    {
+        public int PizzaToppingId { get; set; }
+        public int PizzaId { get; set; }
+        public int ToppingId { get; set; }
+        public int ToppingCount { get; set; }
 
-    //public virtual APizza Pizza { get; set; }
-    //public virtual ATopping Topping { get; set; }
-  }
+        //public virtual APizza Pizza { get; set; }
+        //public virtual ATopping Topping { get; set; }
+    }
 }
